anote_rag/requirements.txt
# Core LangChain (compatible versions)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community

# LLM Providers
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-ollama>=0.1.0

# API clients
anthropic
openai>=1.13.3

# Embeddings
langchain-huggingface
sentence-transformers>=2.2.2

# Vector Database
langchain-chroma
chromadb

# Utilities
python-dotenv==1.0.0